Draw Seven (Pkmn Card Shuffler)
This program I built this for myself, and figured others might benefit from it as well, it's meant to take the input of cards from a deck you have built or want to build, then draws seven random cards as many times as the user requests. This can help if a player's deck currently suffers from a lot of mulligans so they can hopefully readjust and re-get an idea of what their hand will be without shuffling their deck a billion times.